# swagger.yaml
openapi: 3.0.0
info:
  title: API de Webhooks e Integrações
  description: API para receber webhooks da Kiwify, Hotmart, Kirvano e integração com Meta Ads
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Servidor local
paths:
  /webhook/kiwify:
    post:
      summary: Webhook da Kiwify
      description: Endpoint para receber notificações da Kiwify sobre compras, reembolsos e outros eventos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KiwifyWebhook'
      responses:
        '200':
          description: Webhook processado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KiwifyResponse'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Método não permitido
  /webhook/hotmart:
    post:
      summary: Webhook da Hotmart
      description: Endpoint para receber notificações da Hotmart sobre compras, reembolsos e outros eventos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotmartWebhook'
      responses:
        '200':
          description: Webhook processado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotmartResponse'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Método não permitido
  /api/meta-ads:
    post:
      summary: Meta Ads API
      description: |
        Consulta dados do Meta Ads usando um token fornecido no corpo da requisição.
        Retorna métricas como CTR, CAC, investimento total e número de vendas.
        Se o token for inválido ou ocorrer um erro na API, retorna dados simulados como fallback.
      tags:
        - Meta Ads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetaAdsRequest'
            examples:
              tokenExample:
                summary: Exemplo de requisição com token
                value:
                  token: "EAABsbCS1iHgBO..." # Token de exemplo truncado
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
              examples:
                successExample:
                  summary: Exemplo de resposta com sucesso
                  value:
                    success: true
                    message: "Dados do Meta Ads consultados com sucesso"
                    data:
                      ctr: 3.2
                      cac: 12.45
                      investimento: 2500.75
                      numero_vendas: 120
                      campanha_id: "23847239847"
                      campanha_nome: "Campanha de Teste"
                      data_inicio: "2025-02-06"
                      data_fim: "2025-03-06"
                fallbackExample:
                  summary: Exemplo de resposta com dados simulados (fallback)
                  value:
                    success: true
                    message: "Dados simulados do Meta Ads (a API real falhou)"
                    data:
                      ctr: 3.2
                      cac: 12.45
                      investimento: 2500.75
                      numero_vendas: 120
                      campanha_id: "23847239847"
                      campanha_nome: "Campanha de Teste"
                      data_inicio: "2025-02-06"
                      data_fim: "2025-03-06"
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
              example:
                success: false
                message: "Token não fornecido ou formato inválido"
  /meta-ads/metricas:
    post:
      tags:
        - Meta Ads
      summary: Obter métricas do Meta Ads
      description: Obtem métricas como CTR, CAC, investimento total e número de vendas do Meta Ads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MetaAdsRequest'
      responses:
        '200':
          description: Métricas obtidas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
  /meta-ads/campanha/{campaign_id}:
    get:
      tags:
        - Meta Ads
      summary: Obter insights de campanha do Meta Ads
      description: Obtem insights detalhados de uma campanha específica do Meta Ads
      parameters:
        - name: campaign_id
          in: path
          required: true
          schema:
            type: string
          description: ID da campanha
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Token de acesso do Meta Ads
      responses:
        '200':
          description: Insights da campanha obtidos com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
  /meta-ads/conta/{account_id}:
    get:
      tags:
        - Meta Ads
      summary: Obter insights de conta do Meta Ads
      description: Obtem insights detalhados de uma conta de anúncios do Meta Ads
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
          description: ID da conta de anúncios
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Token de acesso do Meta Ads
      responses:
        '200':
          description: Insights da conta obtidos com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaAdsResponse'

components:
  schemas:
    KiwifyWebhook:
      type: object
      required:
        - query
        - body
      properties:
        params:
          type: object
        query:
          type: object
          required:
            - signature
          properties:
            signature:
              type: string
              description: Assinatura para validação do webhook
        body:
          type: object
          required:
            - order_id
            - order_status
          properties:
            order_id:
              type: string
              description: ID único do pedido
            order_ref:
              type: string
              description: Referência do pedido
            order_status:
              type: string
              enum: [paid, rejected, refunded, abandoned]
              description: Status do pedido
            product_type:
              type: string
              description: Tipo do produto
            payment_method:
              type: string
              description: Método de pagamento
            webhook_event_type:
              type: string
              enum: [order_approved, order_rejected, refund_requested, cart_abandoned]
              description: Tipo do evento do webhook
            Product:
              type: object
              properties:
                product_id:
                  type: string
                product_name:
                  type: string
            Customer:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
            Subscription:
              type: object
              properties:
                id:
                  type: string
                status:
                  type: string

    KiwifyResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object

    HotmartWebhook:
      type: object
      required:
        - product
        - purchase
      properties:
        product:
          type: object
          required:
            - ucode
          properties:
            has_co_production:
              type: boolean
            name:
              type: string
            id:
              type: integer
            ucode:
              type: string
        purchase:
          type: object
          required:
            - transaction
            - status
          properties:
            transaction:
              type: string
              description: Código da transação
            status:
              type: string
              enum: [APPROVED, DISPUTE, EXPIRED]
              description: Status da compra
        buyer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string

    HotmartResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string

    MetaAdsRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Token de acesso à API do Meta Ads
          example: "EAABsbCS1iHgBO..."

    MetaAdsResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indica se a operação foi bem-sucedida
        message:
          type: string
          description: Mensagem descritiva sobre o resultado da operação
        data:
          $ref: '#/components/schemas/MetaAdsData'
          description: Dados retornados pela API
        error:
          $ref: '#/components/schemas/ErrorInfo'
          description: Informações detalhadas sobre o erro, se houver

    MetaAdsData:
      type: object
      properties:
        id:
          type: string
          description: ID da campanha ou conta
        nome:
          type: string
          description: Nome da campanha ou conta
        ctr:
          type: number
          description: Taxa de cliques (Click-Through Rate)
        cac:
          type: number
          description: Custo de aquisição de cliente
        investimentoTotal:
          type: number
          description: Valor total investido na campanha
        numeroVendas:
          type: integer
          description: Número de vendas realizadas

    ErrorInfo:
      type: object
      properties:
        type:
          type: string
          description: Tipo do erro
        code:
          type: integer
          description: Código do erro
        message:
          type: string
          description: Mensagem de erro
        details:
          type: string
          description: Detalhes adicionais sobre o erro